Add basis for moving files
