+++ AuroraEditor/Base/NavigatorSidebar/Model/ProjectNavigator/ProjectNavigatorViewController.swift
@@ -7,6 +7,8 @@
 
 import SwiftUI
 
+private let dragType: NSPasteboard.PasteboardType = .fileURL
+
 /// A `NSViewController` that handles the **ProjectNavigator** in the **NavigatorSideabr**.
 ///
 /// Adds a ``outlineView`` inside a ``scrollView`` which shows the folder structure of the
@@ -46,7 +48,7 @@ final class ProjectNavigatorViewController: NSViewController {
     /// This helps determine whether or not to send an `openTab` when the selection changes.
     /// Used b/c the state may update when the selection changes, but we don't necessarily want
     /// to open the file a second time.
-    private var shouldSendSelectionUpdate: Bool = true
+    var shouldSendSelectionUpdate: Bool = true
 
     /// Setup the ``scrollView`` and ``outlineView``
     override func loadView() {
@@ -74,6 +76,7 @@ final class ProjectNavigatorViewController: NSViewController {
         scrollView.hasVerticalScroller = true
         scrollView.hasHorizontalScroller = false
         scrollView.autohidesScrollers = true
+        outlineView.registerForDraggedTypes([dragType])
 
         outlineView.expandItem(outlineView.item(atRow: 0))
         saveExpansionState()
@@ -188,105 +191,6 @@ final class ProjectNavigatorViewController: NSViewController {
         isExpandingThings = false
     }
 
-}
-
-// MARK: - NSOutlineViewDataSource
-
-extension ProjectNavigatorViewController: NSOutlineViewDataSource {
-    func outlineView(_ outlineView: NSOutlineView, numberOfChildrenOfItem item: Any?) -> Int {
-        guard let workspace = self.workspace else { return 0 }
-        if let item = item as? Item {
-            return item.appearanceWithinChildrenOf(searchString: workspace.filter,
-                                                   ignoreDots: true,
-                                                   ignoreTilde: true)
-        }
-        return content.count
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, child index: Int, ofItem item: Any?) -> Any {
-        guard let workspace = self.workspace,
-              let item = item as? Item
-        else { return content[index] }
-
-        return item.childrenSatisfying(searchString: workspace.filter,
-                                       ignoreDots: true,
-                                       ignoreTilde: true)[index]
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, isItemExpandable item: Any) -> Bool {
-        if let item = item as? Item {
-            return item.children != nil
-        }
-        return false
-    }
-}
-
-// MARK: - NSOutlineViewDelegate
-
-extension ProjectNavigatorViewController: NSOutlineViewDelegate {
-    func outlineView(_ outlineView: NSOutlineView,
-                     shouldShowCellExpansionFor tableColumn: NSTableColumn?, item: Any) -> Bool {
-        true
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, shouldShowOutlineCellForItem item: Any) -> Bool {
-        true
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, viewFor tableColumn: NSTableColumn?, item: Any) -> NSView? {
-
-        guard let tableColumn = tableColumn else { return nil }
-
-        let frameRect = NSRect(x: 0, y: 0, width: tableColumn.width, height: rowHeight)
-
-        return ProjectNavigatorTableViewCell(frame: frameRect, item: item as? Item)
-    }
-
-    func outlineViewSelectionDidChange(_ notification: Notification) {
-        guard let outlineView = notification.object as? NSOutlineView else {
-            return
-        }
-
-        let selectedIndex = outlineView.selectedRow
-
-        guard let navigatorItem = outlineView.item(atRow: selectedIndex) as? Item else { return }
-
-        // update the outlineview selection in the workspace. This is used by the bottom toolbar
-        // when the + button is clicked to create a new file.
-        workspace?.newFileModel.outlineViewSelection = navigatorItem
-
-        if !(workspace?.selectionState.openedTabs.contains(navigatorItem.tabID) ?? false) {
-            if navigatorItem.children == nil && shouldSendSelectionUpdate {
-                workspace?.openTab(item: navigatorItem)
-                Log.info("Opened a new tab for: \(navigatorItem.url)")
-            }
-        }
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, heightOfRowByItem item: Any) -> CGFloat {
-        rowHeight // This can be changed to 20 to match Xcode's row height.
-    }
-
-    func outlineViewItemDidExpand(_ notification: Notification) {
-        updateSelection()
-        saveExpansionState()
-    }
-
-    func outlineViewItemDidCollapse(_ notification: Notification) {
-        saveExpansionState()
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, itemForPersistentObject object: Any) -> Any? {
-        guard let id = object as? Item.ID,
-              let item = try? workspace?.fileSystemClient?.getFileItem(id) else { return nil }
-        return item
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, persistentObjectForItem item: Any?) -> Any? {
-        guard let item = item as? Item else { return nil }
-        return item.id
-    }
-
     /// Recursively gets and selects an ``Item`` from an array of ``Item`` and their `children` based on the `id`.
     /// - Parameters:
     ///   - id: the id of the item item
@@ -343,6 +247,73 @@ extension ProjectNavigatorViewController: NSOutlineViewDelegate {
             expandParent(item: parent)
         }
         outlineView.expandItem(item)
+    }
+}
+
+// MARK: - NSOutlineViewDataSource
+
+extension ProjectNavigatorViewController: NSOutlineViewDataSource {
+    func outlineView(_ outlineView: NSOutlineView, numberOfChildrenOfItem item: Any?) -> Int {
+        guard let workspace = self.workspace else { return 0 }
+        if let item = item as? Item {
+            return item.appearanceWithinChildrenOf(searchString: workspace.filter,
+                                                   ignoreDots: true,
+                                                   ignoreTilde: true)
+        }
+        return content.count
+    }
+
+    func outlineView(_ outlineView: NSOutlineView, child index: Int, ofItem item: Any?) -> Any {
+        guard let workspace = self.workspace,
+              let item = item as? Item
+        else { return content[index] }
+
+        return item.childrenSatisfying(searchString: workspace.filter,
+                                       ignoreDots: true,
+                                       ignoreTilde: true)[index]
+    }
+
+    func outlineView(_ outlineView: NSOutlineView, isItemExpandable item: Any) -> Bool {
+        if let item = item as? Item {
+            return item.children != nil
+        }
+        return false
+    }
+
+    // MARK: Drag and Drop
+    func outlineView(_ outlineView: NSOutlineView, pasteboardWriterForItem item: Any) -> NSPasteboardWriting? {
+        guard let fileItem = item as? FileItem else {
+            Log.info("Item is not file item")
+            return nil
+        }
+        let pboarditem = NSPasteboardItem()
+        let success = pboarditem.setString(fileItem.url.path, forType: dragType)
+        return pboarditem
+    }
+
+    func outlineView(_ outlineView: NSOutlineView, validateDrop info: NSDraggingInfo,
+                     proposedItem item: Any?, proposedChildIndex index: Int) -> NSDragOperation {
+        guard let fileItem = item as? FileItem,
+              let draggedString = info.draggingPasteboard.string(forType: dragType)
+        else { return NSDragOperation(arrayLiteral: [])}
+
+        // if the item is being dragged onto an item that it contains, do not move it.
+        // TODO: Fix edge cases like ~/Some not being able to move to ~/Something/ABC
+        if fileItem.isFolder && !fileItem.url.path.hasPrefix(draggedString) {
+            return NSDragOperation.move
+        }
+        return NSDragOperation(arrayLiteral: [])
+    }
+
+    func outlineView(_ outlineView: NSOutlineView, acceptDrop info: NSDraggingInfo,
+                     item: Any?, childIndex index: Int) -> Bool {
+        guard let fileItem = item as? FileItem,
+              let draggedString = info.draggingPasteboard.string(forType: dragType),
+              fileItem.isFolder && !fileItem.url.path.hasPrefix(draggedString)
+        else { return false }
+        let draggedURL = URL(fileURLWithPath: draggedString)
+        Log.info("Drop: \(fileItem.url) onto \(draggedURL)")
+        return true
     }
 }
